requirements.txt
aiogram==3.0.0
gTTS==2.3.2
num2words==0.5.12
openai==0.28.0
playsound==1.3.0
pycairo==1.24.0
pydub==0.25.1
PyGObject==3.46.0
pyTelegramBotAPI==4.13.0
python-dotenv==1.0.0
python-telegram-bot==20.5
revChatGPT==6.8.6
SpeechRecognition==3.10.0